pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "murmur"
version = "0.3.0"
description = "Meeting notes CLI — record, transcribe, summarize"
requires-python = ">=3.10"
dependencies = [
    "openai-whisper",
    "sounddevice",
    "numpy",
    "scipy",
    "click",
    "rich",
    "python-dotenv",
]

[project.optional-dependencies]
# LLM providers
anthropic = ["anthropic"]
openai = ["openai"]
gemini = ["google-generativeai"]
groq = ["groq"]
ollama = ["ollama"]
tui = ["textual>=1.0"]

# New feature groups
clipboard = ["pyperclip"]
export = ["pypandoc", "pypandoc-binary"]
mlx = ["mlx-whisper"]
faster = ["faster-whisper"]
watch = ["psutil"]
diarize = ["pyannote.audio"]

# Bundle (excludes mlx=Apple-only, diarize=heavy PyTorch)
all = [
    "anthropic", "openai", "google-generativeai", "groq", "ollama",
    "textual>=1.0", "pyperclip", "pypandoc", "pypandoc-binary",
    "faster-whisper", "psutil",
]

[tool.setuptools.packages.find]
include = ["murmur*"]

[project.scripts]
murmur = "murmur.cli:cli"
